# fronted/OPENAPI/ExamSystemApi.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ExamSystemApi",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/teachers/{page}/{size}": {
      "get": {
        "summary": "findAll",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "查询所有教师",
                      "data": {
                        "records": [
                          {
                            "teacherId": 0,
                            "teacherName": "",
                            "institute": "",
                            "sex": "",
                            "tel": "",
                            "email": "",
                            "pwd": "",
                            "cardId": "",
                            "type": "",
                            "role": ""
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "ascs": [
                          ""
                        ],
                        "descs": [
                          ""
                        ],
                        "optimizeCountSql": false,
                        "isSearchCount": false,
                        "searchCount": false,
                        "pages": 0,
                        "asc": [
                          ""
                        ],
                        "desc": [
                          ""
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/teacher/{teacherId}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": {
                        "teacherId": 0,
                        "teacherName": "",
                        "institute": "",
                        "sex": "",
                        "tel": "",
                        "email": "",
                        "pwd": "",
                        "cardId": "",
                        "type": "",
                        "role": ""
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/teacher": {
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Teacher",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "add",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Teacher",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Teacher": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "integer",
            "description": ""
          },
          "teacherName": {
            "type": "string",
            "description": ""
          },
          "institute": {
            "type": "string",
            "description": ""
          },
          "sex": {
            "type": "string",
            "description": ""
          },
          "tel": {
            "type": "string",
            "description": ""
          },
          "email": {
            "type": "string",
            "description": ""
          },
          "pwd": {
            "type": "string",
            "description": ""
          },
          "cardId": {
            "type": "string",
            "description": ""
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "role": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ApiResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码，表示一种错误类型\n请求成功，状态码为200"
          },
          "message": {
            "type": "string",
            "description": "对错误代码的详细解释"
          },
          "data": {
            "type": "integer",
            "description": "返回的结果包装在value中，value可以是单个对象"
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": []
}